refactor: move compliance button to compliance related directory
